Enhance system logic

+++ src/main/java/com/code81/library_management/logic/service_impl/CategoryServiceImpl.java
@@ -53,6 +53,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new RuntimeException("Category not found");
+        }
         categoryRepository.deleteById(id);
     }
+
 }
